feat: implement OpenGL extensions loader and initialize game navigation, bot, and voice repacketizer subsystems

+++ src/opus-1.1/src/repacketizer.c
@@ -219,7 +219,7 @@ opus_int32 opus_repacketizer_out_range_impl(OpusRepacketizer *rp, int begin, int
    }
    if (pad)
    {
-      for (i=ptr-data;i<maxlen;i++)
+      for (i=(int)(ptr-data);i<maxlen;i++)
          data[i] = 0;
    }
    return tot_size;
